Add files via upload
In this project, the process are:
First, we print the number based on the position of the byte numbers array, starting from number 0 to 9 and have a delay for 1 second after printing a number.

Then we have a loop function so the numbers will keep printed. For the loop direction, if direction = 1, the numbers will be printed in sequence based on the numbers array. 
If direction = -1, the numbers will be printed backwards (increment or decrement position based on the current loop state).

If the number shown on the 7 Segment LED is 0 or 9, the piezo buzzer will be buzzing three times along with the number shown on the LED will also be blinking three times. After that, the numbers printed will change direction.